objects: domain set, increase timeout
Increase the timeout from 10 to 11 minutes to prevent potential
issues with the cron job.
The cron job runs every 10 minutes, and if it is delayed,
the IP set might be empty for a brief period.
Extending the timeout ensures the IP set remains populated even
if the cron job runs late.